// src/app/models/workposition.ts
/**
 * Đối tượng vị trí công việc
 */
export class WorkPosition {
    // ID: string;
    // PositionName: string;
    // WorkingPositionName: string;
    // WorkTitle: string;
    // WorkPositionDescription;
    // SalaryLevel: number;
    
    PositionID: string;
    PositionCode: string;
    PositionName: string;
    PositionDescription: string;
    SalaryID: string;
    CreatedDate: Date;
    ModifiedDate: Date;
    CreatedBy: string;
}